pyproject.toml
[build-system]
requires = ['setuptools>=42']
build-backend = 'setuptools.build_meta'

[project]
name = "setup-server"
version = "0.1.0"

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
dependencies = [
    'click',
    "filelock",
    "pyyaml"
]

requires-python = ">=3.9"
authors = [{ name = "Shahim Essaid" }]

[project.scripts]
setup-servers = "setupservers.setup_servers:run"
setup-servers-install = "setupservers.setup_servers:install"

[tool.setuptools]
zip-safe = false
include-package-data = true

[project.optional-dependencies]
build = ["build", "twine"]